비디오 미리보기 기능 추가 및 UI 개선
- FileDetailWidget에 비디오 미리보기 기능을 추가하고, 관련 UI 요소를 설정했습니다.
- 비디오 파일 형식에 대한 지원을 추가하고, 재생 및 볼륨 조절 기능을 구현했습니다.
- MainWindow 초기화 시 file_server 인스턴스를 제거했습니다.
- 전체 화면 상태에서 splitter 크기 설정을 정수형으로 변경했습니다.

+++ main_window.py
@@ -66,7 +66,7 @@ class MainWindow(QMainWindow):
         if event.type() == QEvent.WindowStateChange:
             if self.isMaximized():
                 # 전체 화면: file_detail을 80%, file_list를 20%로 설정
-                self.splitter.setSizes([self.height() * 0.8, self.height() * 0.2])
+                self.splitter.setSizes([int(self.height() * 0.8), int(self.height() * 0.2)])
             else:
                 # 일반 화면: 50:50 비율로 복원
                 self.splitter.setSizes([self.height() // 2, self.height() // 2])
